+++ src/test/java/com/zerobase/foodlier/module/dm/dm/service/DmServiceTest.java
@@ -33,13 +33,13 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class DmServiceImplTest {
+public class DmServiceTest {
     @Mock
     private DmRepository dmRepository;
     @Mock
     private DmRoomRepository dmRoomRepository;
     @InjectMocks
-    private DmServiceImpl dmService;
+    private DmService dmService;
 
     @Test
     @DisplayName("채팅작성 - 채팅")
@@ -190,7 +190,7 @@ public class DmServiceImplTest {
                 .willReturn(dmPage);
 
         // when
-        MessageResponseDto messageResponseDto = dmService.getDmList(id, roomId, dmId);
+        MessageResponseDto messageResponseDto = dmService.getDmList(roomId, dmId);
 
         // then
         assertNotNull(messageResponseDto);
@@ -211,7 +211,7 @@ public class DmServiceImplTest {
 
         // when
         DmRoomException dmRoomException = assertThrows(DmRoomException.class,
-                () -> dmService.getDmList(id, roomId, dmId));
+                () -> dmService.getDmList(roomId, dmId));
 
         // then
         assertEquals(DM_ROOM_NOT_FOUND, dmRoomException.getErrorCode());
@@ -241,7 +241,7 @@ public class DmServiceImplTest {
 
         // when
         DmException dmException = assertThrows(DmException.class,
-                () -> dmService.getDmList(id, roomId, dmId));
+                () -> dmService.getDmList(roomId, dmId));
 
         // then
         assertEquals(NO_SUCH_DM, dmException.getErrorCode());
